fixed login and logout with post updates

+++ onthir/main/views.py
@@ -5,6 +5,7 @@ from rest_framework import status
 from .models import Category, Post
 from .serializers import CategorySerializer, PostSerializer
 from django.shortcuts import get_object_or_404
+from rest_framework.permissions import IsAdminUser, IsAuthenticated
 
 # Handle GET and POST operations; they don't require id
 class CategoryListAPIView(APIView):
@@ -14,7 +15,9 @@ class CategoryListAPIView(APIView):
         return Response(serializer.data, status=status.HTTP_200_OK)
     # post request
     def post(self, request):
+
         serializer = CategorySerializer(data=request.data)
+        
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_201_CREATED)
@@ -46,12 +49,24 @@ class CategoryDetailAPIView(APIView):
 class PostListAPIView(APIView):
     # get method
     def get(self, request):
-        posts = Post.objects.all()
+        category_id = request.query_params.get('category_id', None)
+        category_name = request.query_params.get('category_name', None)
+
+        if category_id:
+            posts = Post.objects.filter(category__id=category_id).select_related('category')
+        elif category_name:
+            posts = Post.objects.filter(category__name__iexact=category_name).select_related('category')
+        else:
+            posts = Post.objects.select_related('category').all()
         serializer = PostSerializer(posts, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
     
     # post method
+    
     def post(self, request):
+        
+        if not request.user.is_authenticated:
+            return Response({'error': 'Permission Denied.'}, status=403)
         serializer = PostSerializer(data=request.data)
         print(serializer)
         if serializer.is_valid():
@@ -65,18 +80,26 @@ class PostDetailAPIView(APIView):
         serializer = PostSerializer(post)
         return Response(serializer.data, status=status.HTTP_200_OK)
     
-    # put request
+    # Only logged-in users can update
     def put(self, request, id):
+        print(request)
+        if not request.user.is_authenticated:
+            return Response({"detail": "Authentication required."}, status=status.HTTP_403_FORBIDDEN)
+        
         post = get_object_or_404(Post, id=id)
         serializer = PostSerializer(post, data=request.data, partial=False)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_200_OK)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-    
-    # delete request
+
+    # Only logged-in users can delete
     def delete(self, request, id):
+        print("Request: ", request)
+        print("Status: ", request.user.is_authenticated)
+        if not request.user.is_authenticated:
+            return Response({"detail": "Authentication required."}, status=status.HTTP_403_FORBIDDEN)
+        
         post = get_object_or_404(Post, id=id)
         post.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
-
